requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
h5py==3.10.0
httplib2==0.22.0
idna==3.6
ipykernel==6.29.0
ipython==8.21.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.9
mediapipe-silicon==0.9.2.1
ml-dtypes==0.2.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
stack-data==0.6.3
starlette==0.37.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
termcolor==2.4.0
tflite==2.10.0
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.29.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
